Bug 390735 - Fix places where NSObject id was int in 64-bit code

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/SWTApplicationDelegate.java
@@ -15,7 +15,7 @@ public class SWTApplicationDelegate extends NSObject {
 		super(0);
 	}
 		
-	public SWTApplicationDelegate(int id) {
+	public SWTApplicationDelegate(long /*int*/ id) {
 		super(id);
 	}
 }
