+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/SWTTextView.java
@@ -16,7 +16,7 @@ public SWTTextView() {
 	super(0);
 }
 
-public SWTTextView(int id) {
+public SWTTextView(long /*int*/ id) {
 	super(id);
 }
